requirements.txt
addict==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
backports.zstd==1.3.0
brotli==1.2.0
certifi==2026.2.25
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
dllist==2.0.0
easydict==1.13
exceptiongroup==1.3.1
filelock==3.19.1
flatbuffers==25.12.19
fonttools==4.60.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
humanfriendly==10.0
idna==3.11
inflate64==1.0.4
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
mmcv==2.2.0
mmengine==0.10.7
mpmath==1.3.0
multivolumefile==0.2.3
networkx==3.3
ninja==1.13.0
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.48
nvidia-modelopt==0.41.0
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.8.90
onnx==1.20.1
onnx-ir==0.1.15
onnxruntime-gpu==1.23.2
onnxscript==0.6.0
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.9.4
protobuf==6.33.5
psutil==7.2.2
PuLP==3.3.0
py7zr==1.1.0
pybcj==1.0.7
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pyppmd==1.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
rich==14.3.2
rtree==1.4.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
svg.path==7.0
sympy==1.14.0
tensorboardX==2.6.4
termcolor==3.3.0
texttable==1.7.0
threadpoolctl==3.6.0
timm==1.0.26
tomli==2.4.1
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
trimesh==4.12.1
triton==3.5.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
yapf==0.43.0
zipfile-deflate64==0.2.0
